+++ packages/sdk/src/grpc/translation.ts
@@ -1425,7 +1425,7 @@ export function pendingUpdate(pendingUpdate: GRPC.PendingUpdate): SDK.PendingUpd
     };
 }
 
-export function trPendingUpdateEffect(pendingUpdate: GRPC.PendingUpdate): SDK.PendingUpdateEffect {
+export function trPendingUpdateEffect(pendingUpdate: GRPC.PendingUpdate): Upward<SDK.PendingUpdateEffect> {
     const effect = pendingUpdate.effect;
     switch (effect.oneofKind) {
         case 'protocol':
@@ -1510,13 +1510,17 @@ export function trPendingUpdateEffect(pendingUpdate: GRPC.PendingUpdate): SDK.Pe
                 },
             };
         case undefined:
-            throw Error('Unexpected missing pending update');
+            return null;
     }
 }
 
-function trUpdatePayload(updatePayload: GRPC.UpdatePayload | undefined): SDK.UpdateInstructionPayload {
-    const payload = updatePayload?.payload;
-    switch (payload?.oneofKind) {
+function trUpdatePayload(updatePayload: GRPC.UpdatePayload | undefined): Upward<SDK.UpdateInstructionPayload> {
+    if (updatePayload === undefined) {
+        throw new Error('Unexpected missing update payload');
+    }
+
+    const payload = updatePayload.payload;
+    switch (payload.oneofKind) {
         case 'protocolUpdate':
             return trProtocolUpdate(payload.protocolUpdate);
         case 'electionDifficultyUpdate':
@@ -1594,7 +1598,7 @@ function trUpdatePayload(updatePayload: GRPC.UpdatePayload | undefined): SDK.Upd
                 },
             };
         case undefined:
-            throw new Error('Unexpected missing update payload');
+            return null;
     }
 }
 
@@ -2508,7 +2512,7 @@ function trAccountAmount(
     };
 }
 
-export function blockSpecialEvent(specialEvent: GRPC.BlockSpecialEvent): SDK.BlockSpecialEvent {
+export function blockSpecialEvent(specialEvent: GRPC.BlockSpecialEvent): Upward<SDK.BlockSpecialEvent> {
     const event = specialEvent.event;
     switch (event.oneofKind) {
         case 'bakingRewards': {
@@ -2599,7 +2603,7 @@ export function blockSpecialEvent(specialEvent: GRPC.BlockSpecialEvent): SDK.Blo
             };
         }
         case undefined: {
-            throw Error('Error translating BlockSpecialEvent: unexpected undefined');
+            return null;
         }
     }
 }